Rename cli to spex and add release script
