+++ azure-pipelines.yml
@@ -24,12 +24,13 @@ steps:
 
 - script: |
     npm run build-cli 
-    pkg build/dist/cli/cli.js
+    pkg build/dist/cli/spex.js
   displayName: "Package CLI"
 
 - script: |
-    echo "./cli-linux init --azure-org $AZURE_ORG --azure-project $AZURE_PROJECT --manifest $GITHUB_MANIFEST --github-manifest-username $GITHUB_MANIFEST_USERNAME --storage-account-key $(STORAGE_ACCOUNT_KEY) --storage-account-name $STORAGE_ACCOUNT_NAME --storage-partition-key $STORAGE_PARTITION_KEY --storage-table-name $STORAGE_TABLE_NAME"
-    ./cli-linux init --azure-org $AZURE_ORG --azure-project $AZURE_PROJECT --manifest $GITHUB_MANIFEST --github-manifest-username $GITHUB_MANIFEST_USERNAME --storage-account-key $(STORAGE_ACCOUNT_KEY) --storage-account-name $STORAGE_ACCOUNT_NAME --storage-partition-key $STORAGE_PARTITION_KEY --storage-table-name $STORAGE_TABLE_NAME
+    cd spex-linux
+    echo "./spex init --azure-org $AZURE_ORG --azure-project $AZURE_PROJECT --manifest $GITHUB_MANIFEST --github-manifest-username $GITHUB_MANIFEST_USERNAME --storage-account-key $(STORAGE_ACCOUNT_KEY) --storage-account-name $STORAGE_ACCOUNT_NAME --storage-partition-key $STORAGE_PARTITION_KEY --storage-table-name $STORAGE_TABLE_NAME"
+    ./spex init --azure-org $AZURE_ORG --azure-project $AZURE_PROJECT --manifest $GITHUB_MANIFEST --github-manifest-username $GITHUB_MANIFEST_USERNAME --storage-account-key $(STORAGE_ACCOUNT_KEY) --storage-account-name $STORAGE_ACCOUNT_NAME --storage-partition-key $STORAGE_PARTITION_KEY --storage-table-name $STORAGE_TABLE_NAME
   displayName: "Set Config"
   
 - script: |
