+++ package.json
@@ -21,9 +21,9 @@
     "test": "mocha -r ts-node/register src/tests/Basic.test.ts",
     "private-github-test": "mocha -r ts-node/register src/tests/PrivateGitHub.test.ts",
     "eject": "react-scripts-ts eject",
-    "start-cli": "nodemon --watch 'src/cli/**/*.ts' --exec 'ts-node' src/cli/cli.ts",
+    "start-cli": "nodemon --watch 'src/cli/**/*.ts' --exec 'ts-node' src/cli/spex.ts",
     "create-cli": "npm run build-cli",
-    "build-cli": "rm -rf cli-* && tsc -p . && pkg build/dist/cli/cli.js",
+    "build-cli": "sh release-cli.sh",
     "lint": "tslint -p tsconfig.json -c tslint.json src/**/*.ts",
     "refresh": "rm -rf ./node_modules ./package-lock.json && npm install"
   },
