+++ src/cli/spex.ts
